+++ src/Validator/Object/PropertyDependentValidator.php
@@ -4,6 +4,7 @@ namespace ExtendsSoftware\ExaPHP\Validator\Object;
 
 use ExtendsSoftware\ExaPHP\Validator\AbstractValidator;
 use ExtendsSoftware\ExaPHP\Validator\Exception\TemplateNotFound;
+use ExtendsSoftware\ExaPHP\Validator\Object\Properties\Property;
 use ExtendsSoftware\ExaPHP\Validator\Result\ResultInterface;
 use ExtendsSoftware\ExaPHP\Validator\ValidatorInterface;
 
@@ -24,6 +25,13 @@ class PropertyDependentValidator extends AbstractValidator
     public const PROPERTY_MISSING = 'propertyMissing';
 
     /**
+     * Properties.
+     *
+     * @var array<string, Property>
+     */
+    private array $properties = [];
+
+    /**
      * @param string                                 $property
      * @param array<string, ValidatorInterface>|null $validators
      * @param bool|null                              $strict
@@ -33,7 +41,6 @@ class PropertyDependentValidator extends AbstractValidator
         private ?array          $validators = null,
         private ?bool           $strict = null
     ) {
-        $this->validators ??= [];
         $this->strict ??= true;
 
         foreach ($validators ?? [] as $value => $validator) {